requirements.txt
discord.py==1.7.3
fastapi==0.83.0
uvicorn==0.16.0
python-dotenv==0.20.0
asyncio==3.4.3
schedule==1.1.0